Attempt to change the color of each div
As the mouse passes through, the color of each div should change based
on the mouse entering a new div. I think the problem might be that the
divs I created can't be access from outside the for loop. Maybe need to
find a way to access them from outside the loop?