Create a button that generates a new grid
I added an event listener to the button that would take the prompt and
have that number be the amount of times that the for loop iterated. A
prompt of 40 would produce 40 squares.

The real trouble was getting the grid to reset itself before taking a
new prompt. The initial behavior would be the squares would just add on
to what was already there. I fixed this by adding another event listener
before the one that produces squares.

This event listener would check if there was an element on the page
that was made by the other event listener. If true, it would run a while
loop that removed all of it's children.